requirements.txt
click>=8.1.0
pyopencl
base58
PyNaCl
numpy==1.26.4
mnemonic>=0.20